+++ google-cloud-storage/src/test/java/com/google/cloud/storage/V4PostPolicyTest.java
@@ -16,24 +16,22 @@
 
 package com.google.cloud.storage;
 
-import static com.google.cloud.storage.Storage.V4ConditionType.CONTENT_LENGTH_RANGE;
 import static com.google.cloud.storage.Storage.V4ConditionType.MATCHES;
 import static com.google.cloud.storage.Storage.V4ConditionType.STARTS_WITH;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-import com.google.api.Http;
-import com.google.api.client.http.HttpResponse;
 import com.google.api.core.ApiClock;
 import com.google.auth.oauth2.ServiceAccountCredentials;
-import com.google.cloud.conformance.storage.v1.ConditionalMatches;
+import com.google.cloud.conformance.storage.v1.PolicyConditions;
 import com.google.cloud.conformance.storage.v1.PolicyInput;
 import com.google.cloud.conformance.storage.v1.PostPolicyV4Test;
-import com.google.cloud.conformance.storage.v1.SigningV4Test;
 import com.google.cloud.conformance.storage.v1.TestFile;
-import com.google.cloud.storage.Storage.SignUrlOption;
+import com.google.cloud.conformance.storage.v1.UrlStyle;
 import com.google.cloud.storage.testing.RemoteStorageHelper;
 import com.google.common.base.Charsets;
+import com.google.common.base.Strings;
+import com.google.common.io.BaseEncoding;
 import com.google.protobuf.Timestamp;
 import com.google.protobuf.util.JsonFormat;
 import java.io.IOException;
@@ -42,6 +40,7 @@ import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
 import org.junit.Rule;
@@ -54,124 +53,150 @@ import org.junit.runners.Parameterized.Parameters;
 @RunWith(Parameterized.class)
 public class V4PostPolicyTest {
 
-    private static final String SERVICE_ACCOUNT_JSON_RESOURCE =
-            "com/google/cloud/conformance/storage/v1/test_service_account.not-a-test.json";
-    private static final String TEST_DATA_JSON_RESOURCE =
-            "com/google/cloud/conformance/storage/v1/v4_signatures.json";
+  private static final String SERVICE_ACCOUNT_JSON_RESOURCE =
+      "com/google/cloud/conformance/storage/v1/test_service_account.not-a-test.json";
+  private static final String TEST_DATA_JSON_RESOURCE =
+      "com/google/cloud/conformance/storage/v1/v4_signatures.json";
 
-    private static class FakeClock implements ApiClock {
-        private final AtomicLong currentNanoTime;
+  private static class FakeClock implements ApiClock {
+    private final AtomicLong currentNanoTime;
 
-        public FakeClock(Timestamp timestamp) {
-            this.currentNanoTime =
-                    new AtomicLong(
-                            TimeUnit.NANOSECONDS.convert(timestamp.getSeconds(), TimeUnit.SECONDS)
-                                    + timestamp.getNanos());
-        }
-
-        public long nanoTime() {
-            return this.currentNanoTime.get();
-        }
-
-        public long millisTime() {
-            return TimeUnit.MILLISECONDS.convert(this.nanoTime(), TimeUnit.NANOSECONDS);
-        }
+    public FakeClock(Timestamp timestamp) {
+      this.currentNanoTime =
+          new AtomicLong(
+              TimeUnit.NANOSECONDS.convert(timestamp.getSeconds(), TimeUnit.SECONDS)
+                  + timestamp.getNanos());
     }
 
-    @Rule public TestName testName = new TestName();
-
-    private final PostPolicyV4Test testData;
-    private final ServiceAccountCredentials serviceAccountCredentials;
-
-    /**
-     * @param testData The serialized test data representing the test case.
-     * @param serviceAccountCredentials The credentials to use in this test.
-     * @param description Not used by the test, but used by the parameterized test runner as the name
-     *     of the test.
-     */
-    public V4PostPolicyTest(
-            PostPolicyV4Test testData,
-            ServiceAccountCredentials serviceAccountCredentials,
-            @SuppressWarnings("unused") String description) {
-        this.testData = testData;
-        this.serviceAccountCredentials = serviceAccountCredentials;
+    public long nanoTime() {
+      return this.currentNanoTime.get();
     }
 
-    @Test
-    public void test() {
-        Storage storage =
-                RemoteStorageHelper.create()
-                        .getOptions()
-                        .toBuilder()
-                        .setCredentials(serviceAccountCredentials)
-                        .setClock(new FakeClock(testData.getPolicyInput().getTimestamp()))
-                        .build()
-                        .getService();
+    public long millisTime() {
+      return TimeUnit.MILLISECONDS.convert(this.nanoTime(), TimeUnit.NANOSECONDS);
+    }
+  }
 
-        BlobInfo blob = BlobInfo.newBuilder(testData.getPolicyInput().getBucket(), testData.getPolicyInput().getObject()).build();
+  @Rule public TestName testName = new TestName();
 
-        PolicyInput policyInput = testData.getPolicyInput();
-        Storage.V4PostConditions.Builder builder = Storage.V4PostConditions.newBuilder();
-        System.out.println(testData.getPolicyOutput().getExpectedDecodedPolicy());
+  private final PostPolicyV4Test testData;
+  private final ServiceAccountCredentials serviceAccountCredentials;
 
-        for(ConditionalMatches match : policyInput.getConditions().getMatchesList()) {
-            Storage.V4ConditionType type = STARTS_WITH;
-            switch(match.getExpression(0)) {
-                case "startsWith":
-                    type = STARTS_WITH;
-                    break;
-                case "matches":
-                    type = MATCHES;
-                    break;
-                case "content-length-range":
-                    type = CONTENT_LENGTH_RANGE;
-                    break;
-            }
-            String key = match.getExpression(1).replace("$", "");
-            String value = match.getExpression(2);
-            switch(key) {
-                case "acl":
-                    builder.addAclCondition(type, value);
-            }
-        }
+  /**
+   * @param testData The serialized test data representing the test case.
+   * @param serviceAccountCredentials The credentials to use in this test.
+   * @param description Not used by the test, but used by the parameterized test runner as the name
+   *     of the test.
+   */
+  public V4PostPolicyTest(
+      PostPolicyV4Test testData,
+      ServiceAccountCredentials serviceAccountCredentials,
+      @SuppressWarnings("unused") String description) {
+    this.testData = testData;
+    this.serviceAccountCredentials = serviceAccountCredentials;
+  }
 
-        storage.generateV4PresignedPostPolicy(null, blob, null, builder.build(), testData.getPolicyInput().getExpiration());
-        System.out.println(testData.getPolicyOutput().getFieldsMap().get("policy"));
+  @Test
+  public void test() {
+    Storage storage =
+        RemoteStorageHelper.create()
+            .getOptions()
+            .toBuilder()
+            .setCredentials(serviceAccountCredentials)
+            .setClock(new FakeClock(testData.getPolicyInput().getTimestamp()))
+            .build()
+            .getService();
+
+    BlobInfo blob =
+        BlobInfo.newBuilder(
+                testData.getPolicyInput().getBucket(), testData.getPolicyInput().getObject())
+            .build();
+
+    PolicyInput policyInput = testData.getPolicyInput();
+    Storage.V4PostConditions.Builder builder = Storage.V4PostConditions.newBuilder();
+
+    Map<String, String> fields = policyInput.getFieldsMap();
+
+    PolicyConditions conditions = policyInput.getConditions();
+
+    if (!Strings.isNullOrEmpty(fields.get("success_action_redirect"))) {
+      builder.addSuccessActionRedirectUrlCondition(MATCHES, fields.get("success_action_redirect"));
     }
 
-    /**
-     * Load all of the tests and return a {@code Collection<Object[]>} representing the set of tests.
-     * Each entry in the returned collection is the set of parameters to the constructor of this test
-     * class.
-     *
-     * <p>The results of this method will then be run by JUnit's Parameterized test runner
-     */
-    @Parameters(name = "{2}")
-    public static Collection<Object[]> testCases() throws IOException {
-        ClassLoader cl = Thread.currentThread().getContextClassLoader();
-
-        InputStream credentialsStream = cl.getResourceAsStream(SERVICE_ACCOUNT_JSON_RESOURCE);
-        assertNotNull(
-                String.format("Unable to load service account json: %s", SERVICE_ACCOUNT_JSON_RESOURCE),
-                credentialsStream);
-
-        InputStream dataJson = cl.getResourceAsStream(TEST_DATA_JSON_RESOURCE);
-        assertNotNull(
-                String.format("Unable to load test definition: %s", TEST_DATA_JSON_RESOURCE), dataJson);
-
-        ServiceAccountCredentials serviceAccountCredentials =
-                ServiceAccountCredentials.fromStream(credentialsStream);
-
-        InputStreamReader reader = new InputStreamReader(dataJson, Charsets.UTF_8);
-        TestFile.Builder testBuilder = TestFile.newBuilder();
-        JsonFormat.parser().merge(reader, testBuilder);
-        TestFile testFile = testBuilder.build();
-
-        List<PostPolicyV4Test> tests = testFile.getPostPolicyV4TestsList();
-        ArrayList<Object[]> data = new ArrayList<>(tests.size());
-        for (PostPolicyV4Test test : tests) {
-            data.add(new Object[] {test, serviceAccountCredentials, test.getDescription()});
-        }
-        return data;
+    if (!Strings.isNullOrEmpty(fields.get("success_action_status"))) {
+      builder.addSuccessActionStatusCondition(
+          MATCHES, Integer.parseInt(fields.get("success_action_status")));
     }
+
+    if (conditions != null) {
+      if (!conditions.getStartsWithList().isEmpty()) {
+        builder.addCustomCondition(
+            STARTS_WITH, conditions.getStartsWith(0).replace("$", ""), conditions.getStartsWith(1));
+      }
+      if (!conditions.getContentLengthRangeList().isEmpty()) {
+        builder.addContentLengthRange(
+            conditions.getContentLengthRange(0), conditions.getContentLengthRange(1));
+      }
+    }
+
+    Storage.V4PostFields v4Fields = Storage.V4PostFields.of(fields);
+
+    Storage.V4PostPolicyOption style = Storage.V4PostPolicyOption.withPathStyle();
+
+    if (policyInput.getUrlStyle().equals(UrlStyle.VIRTUAL_HOSTED_STYLE)) {
+      style = Storage.V4PostPolicyOption.withVirtualHostedStyle();
+    } else if (policyInput.getUrlStyle().equals(UrlStyle.PATH_STYLE)) {
+      style = Storage.V4PostPolicyOption.withPathStyle();
+    } else if (policyInput.getUrlStyle().equals(UrlStyle.BUCKET_BOUND_HOSTNAME)) {
+      style =
+          Storage.V4PostPolicyOption.withBucketBoundHostname(
+              policyInput.getBucketBoundHostname(),
+              Storage.UriScheme.valueOf(policyInput.getScheme().toUpperCase()));
+    }
+
+    Storage.V4PostPolicy policy =
+        storage.generateV4PresignedPostPolicy(
+            blob, v4Fields, builder.build(), testData.getPolicyInput().getExpiration(), style);
+    assertEquals(testData.getPolicyOutput().getFieldsMap(), policy.getFields());
+    assertEquals(
+        testData.getPolicyOutput().getExpectedDecodedPolicy(),
+        new String(BaseEncoding.base64().decode(policy.getFields().get("policy"))));
+    assertEquals(testData.getPolicyOutput().getUrl(), policy.getUrl());
+  }
+
+  /**
+   * Load all of the tests and return a {@code Collection<Object[]>} representing the set of tests.
+   * Each entry in the returned collection is the set of parameters to the constructor of this test
+   * class.
+   *
+   * <p>The results of this method will then be run by JUnit's Parameterized test runner
+   */
+  @Parameters(name = "{2}")
+  public static Collection<Object[]> testCases() throws IOException {
+    ClassLoader cl = Thread.currentThread().getContextClassLoader();
+
+    InputStream credentialsStream = cl.getResourceAsStream(SERVICE_ACCOUNT_JSON_RESOURCE);
+    assertNotNull(
+        String.format("Unable to load service account json: %s", SERVICE_ACCOUNT_JSON_RESOURCE),
+        credentialsStream);
+
+    InputStream dataJson = cl.getResourceAsStream(TEST_DATA_JSON_RESOURCE);
+    assertNotNull(
+        String.format("Unable to load test definition: %s", TEST_DATA_JSON_RESOURCE), dataJson);
+
+    ServiceAccountCredentials serviceAccountCredentials =
+        ServiceAccountCredentials.fromStream(credentialsStream);
+
+    InputStreamReader reader = new InputStreamReader(dataJson, Charsets.UTF_8);
+    TestFile.Builder testBuilder = TestFile.newBuilder();
+    JsonFormat.parser().merge(reader, testBuilder);
+    TestFile testFile = testBuilder.build();
+
+    List<PostPolicyV4Test> tests = testFile.getPostPolicyV4TestsList();
+    ArrayList<Object[]> data = new ArrayList<>(tests.size());
+    for (PostPolicyV4Test test : tests) {
+      data.add(new Object[] {test, serviceAccountCredentials, test.getDescription()});
+    }
+    return data;
+  }
 }